requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
click==7.1.2
dj-database-url==0.5.0
Django==3.0.7
djangorestframework==3.11.0
gunicorn==20.0.4
h11==0.9.0
numpy==1.18.5
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
prettyconf==2.1.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
uvicorn==0.11.5
websockets==8.1
wincertstore==0.2
